Enable floats in the map matrix

+++ functions/init_funcs/initMap.py
@@ -11,7 +11,7 @@ def init_map(map_name):
     """
 
     img = Image.open("data/img/{}".format(map_name)).convert("RGB")
-    arr3d = np.array(img)
+    arr3d = np.array(img, dtype='float32')
     arr2d = arr3d[:, :, 0]
 
     arr2d[arr2d == 255] = 1
